Fix test failures on 32-bit systems due to integer overflow. Note that size_t is a 32-bit type on 32-bit Linux systems.

+++ tests/test-bmat8.cpp
@@ -589,7 +589,7 @@ namespace libsemigroups {
     BMat8 zero(0);
     REQUIRE(BMat8({{0, 0}, {0, 0}}) == zero);
     REQUIRE(BMat8({{0, 0}, {0, 1}}) != zero);
-    REQUIRE(BMat8({{0, 0}, {0, 1}}) == BMat8(size_t(1) << 54));
+    REQUIRE(BMat8({{0, 0}, {0, 1}}) == BMat8(uint64_t(1) << 54));
 
     REQUIRE_THROWS_AS(BMat8({{0, 0}}), LibsemigroupsException);
     REQUIRE_THROWS_AS(BMat8({{0, 1}}), LibsemigroupsException);
